Extract rgba imagesource
